Move choropleth to its own class
